Rename to Places Manager
- in this commit, only changes that do not affect
  external services (assume URL is the same,
  assume repo is the same, etc.)
- only URL to be updated is the one in the docs.
