+++ config/environments/production.rb
@@ -52,7 +52,7 @@ Rails.application.configure do
 
   # Use a real queuing backend for Active Job (and separate queues per environment).
   # config.active_job.queue_adapter     = :resque
-  # config.active_job.queue_name_prefix = "imminence_production"
+  # config.active_job.queue_name_prefix = "places_manager_production"
 
   config.action_mailer.perform_caching = false
