+++ config/locales/en.yml
@@ -30,6 +30,7 @@
 # available at https://guides.rubyonrails.org/i18n.html.
 
 en:
+  product_name: Places Manager
   presenters:
     data_set:
       status:
